Add react-router-dom to dependencies and refactor views
This commit adds the `react-router-dom` package as a dependency in the `package.json` file. It also refactors the `AppView.tsx` and `index.tsx` files to use the new router. The `RootView.tsx` file was deleted since it is no longer needed. Finally, there were some changes made to import statements in the code.